Rqd shutdown and nimby pynput as default
The shutdown process was changed to cleanly stop the process when shutdown-now is requested.
Pynput is now the default mode for nimby checks, and the factory defaults to NOP if the library
is not installed on the environment.

(cherry picked from commit 3389d0f219d599d336987ac589c9be4f0b945783 on gitlab)